fix: resolve all clippy warnings (-D warnings)
- fp.rs: collapse else-if chains; add #[allow(too_many_arguments)] on
  exec_arith (10 params required by RISC-V FP dispatch)
- pseudo.rs: use RangeInclusive::contains for -2048..=2047 check
- engine.rs: remove redundant u32->u32 cast on csr_addr
- parser.rs: #[allow(dead_code)] on Statement (Span fields kept for
  future error reporting); remove unneeded return statement
- codegen.rs: #[allow(dead_code)] on symbols field (reserved for future
  symbol-table panel)
- symbol_table.rs: #[allow(dead_code)] on all() (future use)
- csr.rs: #[allow(dead_code)] on addr module (M-mode CSRs reserved)
- error.rs: #[allow(dead_code)] on OarsError (Lex/Parse/Assemble
  variants used via anyhow context, not direct construction)
- gui/app.rs: iterate REG_NAMES/FP_REG_NAMES with enumerate() instead
  of indexing by loop variable

+++ src/simulator/engine.rs
@@ -202,7 +202,7 @@ fn exec_csr(
     let f3 = f::funct3(word);
     let rd = f::rd(word);
     let rs1 = f::rs1(word);
-    let csr_addr = (word >> 20) as u32; // bits 31:20
+    let csr_addr = word >> 20; // bits 31:20
     let uimm = rs1 as u32; // for csrrwi/csrrsi/csrrci, rs1 field = uimm
 
     match f3 {
